Keep copy of parsed JSON fields (fix #18)
Until now, jftui tracked parsed JSON fields via pointers referencing the
raw payload directly. This was very fast, but worked only thanks to an
input invariant we lost with JF 10.7.0.
When it finds an escaped string, YAJL uses an internal buffer to store
the decoded string and presents that as callback argument. The buffer is
reused between tokens, reallocating only if it's not long enough.
Clearly, this reuse invalidates any pointer referencing the previous
contents, which is a problem for us if we need more than one escaped
string field for the same item. This happens frequently with non-English
content (think a song title and artist).
Until 10.7.0, JF never used escaped strings, embedding a UTF-8 payload
directly. Starting from 10.7.0, JF escapes all non-ASCII codepoints,
triggering the mentioned YAJL behaviour.

The solution employed is to copy all parsed tokens from YAJL into an
internal jftui buffer. This is done in a way that requires very few
allocations, but still a lot of memory copying. Alas.

The upside is that we've replaced the old copy_buffer (the one
accounting for tokens spanning multiple HTTP chunks) with a much simpler
and more robust mechanism.