Don't clutter config dir with cache files.
mpv 0.36/libmpv 2.1 introduce the rather disagreeable convention of using
the config-dir property to also override the 3D LUT and GLSL shader
cache directories. The result is an unsightly and almost certainly unwanted
mess in the config-dir, so we have to perform the circus act of
reconstructing the most likely MPV cache directory and setting individual
cache directory-related properties to point there.