Replace item count with date limit in NextUp
There's not really a good reason to keep the item count limit in the NextUp
query, except for when you're passing EnableRewatching=true, which however
seems to be badly broken and will make a big mess of things.

(brokenness 1: I have an episode show up, I watch it, it still shows up)
(brokenness 2: I have a series where two episodes show up at the same time)

We will not be passing EnableRewatching=true.

It's however unreasonable to not put any limit at all (not for a technical
limit on either jftui nor Jellyfin, but for usability), so we will go by
date and filter NextUp for the last solar year only. This is hardcoded for
now but one day we might make it configurable.

Fix #39.