Properly open JF playback session
You're supposed to make a POST to /sessions/playing/playing when starting playback
to properly open a session on the server. However, a session is also implicitly
opened if you just start by POST'ing to /sessions/playing/progress (the endpoint
you use to mark watch progress within an item) with a progress marker.
Up until now, this implicit session was started with the timestamp of the request.
With 10.10.7 this changed and the implicit session is started with a timestamp of 0.

This makes it so that the server thinks you last watched your episode or movie right
when ole J.C. had his first wail.

This doesn't play nice with sorting your NextUp section by time watching, especially
when you have a cutoff either by "most recent N items" or by "items more recent than
X date".

jftui was lazily only doing the /sessions/playing/progress updates and it broke.
We now do the tidy thing.

Fix #46