Fix resume prompt in playlist skip
We have to halt playback and all input when we ask the user something.
This was currently not done at all on a playlist skip: the item kept
playing while we tried to prompt the user. mpv was also still accepting
input.

Unfortunately, I've found no reliable way to halt things except to
completely abort the mpv core and restart it. Apparently the
`input-terminal` and `input-vo-keyboard` properties cannot be changed
at runtime.